Add option to start fetcher from last fetched block

+++ src/Api/Types.hs
@@ -12,6 +12,7 @@ module Api.Types (
 ) where
 
 import Data.Aeson (FromJSON, ToJSON)
+import Data.Int (Int64)
 import Data.Text (Text)
 import Data.Vector (Vector)
 import Servant.API.Generic (Generic)
@@ -19,7 +20,7 @@ import Servant.API.Generic (Generic)
 import PlutusData qualified
 
 data FirstFetchBlock = FirstFetchBlock
-    { blockSlot :: Integer
+    { blockSlot :: Int64
     , blockId :: Text
     }
 
@@ -47,7 +48,7 @@ newtype GetDatumsByHashesResponse = GetDatumsByHashesResponse
     deriving anyclass (ToJSON)
 
 data StartBlockFetchingRequest = StartBlockFetchingRequest
-    { slot :: Integer
+    { slot :: Int64
     , id :: Text
     }
     deriving stock (Generic)
